requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1
flask-session==0.5.0
fastapi==0.109.0
uvicorn==0.27.0
PyJWT==2.8.0
pydantic[email]>=2.0.0
requests==2.31.0
slowapi==0.1.9
pytest==8.1.1